Simplify and rename parallelized autoresearch tutorial.
Drop the structured and single-threaded code-edit agent entry points,
remove unused helpers, and rename mle_autoresearch_fanout to
parallelized_autoresearch with parallelized_autoresearch.py as the entry script.
